Restructure documentation and optimize URL slugs
Remove the "basics" prefix from the URL which simplifies the URL structure and also follows the intent to restructure the documentation in general to make it better understandable and only contain documentation pages which actually have contents.
The empty documentation pages are more like a "TODO" list but are not really helpful to anyone reading the docs.